requirements.txt
# for yaml
PyYAML
# for documentation
sphinx
sphinx_bootstrap_theme
